fix: ensure the new date is always between max and min dates
